Change everything to long doubles, optimised EGT function
get now ignores when payoff diff is <= 0

+++ 74100/74100/Agent.h
@@ -15,9 +15,8 @@
 class Agent{
     
 public:
-    
-    float payoff = 0.0f;
-    float fitness = 0.0f;
+    long double payoff = 0.0f;
+    long double fitness = 0.0f;
     int tag = 0;
     std::vector<int> strategy; //index for each tag, 0 means cooperation, 1 means defection to that tag
     std::vector<int> neighbors; //index of his neighbors
@@ -31,8 +30,8 @@ public:
     //TODO: finish this, set things to private..
     /**
      int getTag();
-     float getPayoff();
-     float getFitness();
+     long double getPayoff();
+     long double getFitness();
     //*/
     
 private:
